+++ src/veil/backend/database/oracle_driver/oracle_driver_installer.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals, print_function, division
 from veil.profile.installer import *
 
 
-ORACLE_HOME = DEPENDENCY_DIR / 'instantclient_12_1'
+ORACLE_HOME = DEPENDENCY_INSTALL_DIR / 'instantclient_12_1'
 ORACLE_DRIVER_CONF_PATH = '/etc/ld.so.conf.d/oracle_driver.conf'
 RESOURCE_KEY = 'veil.backend.database.oracle_driver.oracle_driver_resource'
 RESOURCE_VERSION = '12.1.0.1.0'
@@ -28,11 +28,13 @@ def download_oracle_instantclient():
     if os.path.exists(ORACLE_HOME):
         return
     basic_local_path = DEPENDENCY_DIR / 'instantclient-basic-linux.x64-12.1.0.1.0.zip'
-    shell_execute('wget -c {}/instantclient-basic-linux.x64-12.1.0.1.0.zip -O {}'.format(DEPENDENCY_URL, basic_local_path))
-    shell_execute('unzip {} -d {}'.format(basic_local_path, DEPENDENCY_DIR))
+    if not os.path.exists(basic_local_path):
+        shell_execute('wget -c {}/instantclient-basic-linux.x64-12.1.0.1.0.zip -O {}'.format(DEPENDENCY_URL, basic_local_path))
+    shell_execute('unzip {} -d {}'.format(basic_local_path, DEPENDENCY_INSTALL_DIR))
     sdk_local_path = DEPENDENCY_DIR / 'instantclient-sdk-linux.x64-12.1.0.1.0.zip'
-    shell_execute('wget -c {}/instantclient-sdk-linux.x64-12.1.0.1.0.zip -O {}'.format(DEPENDENCY_URL, sdk_local_path))
-    shell_execute('unzip {} -d {}'.format(sdk_local_path, DEPENDENCY_DIR))
+    if not os.path.exists(sdk_local_path):
+        shell_execute('wget -c {}/instantclient-sdk-linux.x64-12.1.0.1.0.zip -O {}'.format(DEPENDENCY_URL, sdk_local_path))
+    shell_execute('unzip {} -d {}'.format(sdk_local_path, DEPENDENCY_INSTALL_DIR))
 
 
 def is_oracle_instantclient_installed():
